Split connection helpers into matching files
